CP-10535: Fix device overheating (#2790)
