+++ packages/k2-alpine/src/hooks/useMotion.ts
@@ -1,5 +1,3 @@
-import { useEffect, useMemo, useState } from 'react'
-import { AppState, Platform } from 'react-native'
 import {
   SharedValue,
   useAnimatedSensor,
@@ -8,30 +6,11 @@ import {
   Value3D
 } from 'react-native-reanimated'
 
-export const useMotion = (): Motion | undefined => {
-  const [appState, setAppState] = useState(AppState.currentState)
-  const shouldAnimate = useMemo(() => appState === 'active', [appState])
-  const rotation = useAnimatedSensor(SensorType.ROTATION)
-  const accelerometer = useAnimatedSensor(SensorType.ACCELEROMETER)
+export const useMotion = (isActive: boolean): Motion | undefined => {
+  const rotation = useAnimatedSensor(SensorType.ROTATION, isActive)
+  const accelerometer = useAnimatedSensor(SensorType.ACCELEROMETER, isActive)
 
-  useEffect(() => {
-    const subscription = AppState.addEventListener('change', nextAppState => {
-      setAppState(nextAppState)
-    })
-
-    return () => {
-      subscription.remove()
-    }
-  }, [])
-
-  useEffect(() => {
-    if (Platform.OS === 'android') {
-      rotation.unregister()
-      accelerometer.unregister()
-    }
-  }, [accelerometer, rotation])
-
-  return shouldAnimate && Platform.OS !== 'android'
+  return isActive
     ? {
         rotation: rotation.sensor,
         accelerometer: accelerometer.sensor
